requirements.txt
ipykernel
langchain
python-dotenv
langchain_community
langchain-text-splitters
langchain_chroma
langchain_groq
streamlit
SpeechRecognition
pyaudio
gtts
pydub
